\name{camtrap-package}
\alias{camtrap-package}
\alias{camtrap}
\docType{package}
\title{
\packageTitle{camtrap}
}
\description{
\packageDescription{camtrap}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{camtrap}
\packageIndices{camtrap}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{camtrap}

Maintainer: \packageMaintainer{camtrap}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
