% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapCross.R
\name{overlapCross}
\alias{overlapCross}
\title{Find crossings of estimated densities}
\usage{
overlapCross(A, B, adjust = 0.8, kmax = 3, n.grid = 128, scale24 = T)
}
\arguments{
\item{A, B}{Numeric vectors of sighting times in radians, i.e. [0,2pi]}

\item{adjust}{Scaling factor for the bandwidth estimate, 
see \code{\link{overlapEst}}}

\item{kmax}{Parameter passed to \code{getBandWidth}, see \code{overlapEst}.}

\item{n.grid}{The number of grid points used to find rough crossing windows}

\item{scale24}{Logical, if true roots are multiplied by 12/pi for 24hr scale}
}
\value{
A vector of crossings times on the requested scale.
}
\description{
\code{overlapCross} Fits densities to provided data and returns the 
locations where they cross.
}
\details{
This function is designed for use with relatively smooth densities with only
a few crossings. It first estimates densities using the provided options
and data. It then finds intervals where crossings occur using a grid search 
with \code{n.grid} points over [0,2pi] before calling \code{\link{uniroot}}
to numerically approximate the crossings in each interval using the
difference between densities.
}
