% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapEstWindow.R
\name{overlapEstWindow}
\alias{overlapEstWindow}
\title{Estimate overlap and excess for a fixed window.}
\usage{
overlapEstWindow(A, B, t0 = 0, t1 = 24, adjust = 0.8, kmax = 3)
}
\arguments{
\item{A, B}{Numeric vectors of sighting times in [0,2pi]}

\item{t0, t1}{The window over which to estimate the overlap and excess stats}

\item{adjust, kmax}{See \code{\link{overlapEst}}.}
}
\value{
A list containing: \describe{
\item{overlap}{The area under the minimum of the two densities within
the window requested using \code{t0} and \code{t1}.}
\item{excess}{Contains the area between the density curves within 
the window. If the functions cross within the window, this is the net
difference. Differencs are for \eqn{f_A} less \eqn{f_B}.}
\item{window}{The requested interval.}
}
}
\description{
Estimates densities for two sets of sighting times and then
returns both the overlap and difference within a specified window.
}
\details{
This functions uses t\code{\link{getBandwidth}} and
\code{\link{densityFit}} to estimate density functions for the
times provided as \code{A} and \code{B}. The various overlap estimates 
described in \code{\link{overlapEst}} are forgone in favor of 
adapative quadarature using \code{\link{integrate}}. The arguments
\code{t0} and \code{t1} are transformed to radian cscale and then
used as the limits of integration, wrapped when appropriate.
}
\author{
James Henderson based on work by
}
