% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapPerm.R
\name{overlapPerm}
\alias{overlapPerm}
\title{Overlap window permutation tests.}
\usage{
overlapPerm(A, B, t0 = 0, t1 = 24, adjust = 0.8, nperm = 1000,
  kmax = 3, parametric = T, two.sided = T, overlapRef = 1,
  excessRef = 0)
}
\arguments{
\item{A}{Numeric vectors of sighting times in [0,2pi]}

\item{B}{Numeric vectors of sighting times in [0,2pi]}

\item{t0}{The window over which to estimate the overlap and excess stats}

\item{t1}{The window over which to estimate the overlap and excess stats}

\item{adjust}{See \code{\link{overlapEst}}.}

\item{nperm}{The number of simulated samples or permutations to generate.}

\item{kmax}{See \code{\link{overlapEst}}.}

\item{two.sided}{If \code{TRUE} the p-values returned for the excess are 
two-sided.P-values for the overlap are always one-sided.}

\item{overlapRef}{The reference value against which to compare the
observed overlap. The default is 1, which makes sense when comparing subsets
of the same species.}

\item{excessRef}{Like \code{overlapRef}, but for the excess.  The default 
value is zero, which should make sense in most cases.}
}
\value{
An object of class overlapPermObj containing a table of values and
information about the function call.
}
\description{
Assess the signficance of the overlap or excess in a window using a
simulation based permutation test.
}
\details{
This function estimates the probability of observing an
overlap/excess in a given interval as extreme as the one observed under the
null hypothesis that A and B actually come from the same distribution. By
default, this hypothesis is tested by simulating new data A' and B' from a
density estimated by concatenating A and B.  Setting \code{parametric = F} 
will instead give a tradiational permutation test (not yet implemented.)
}
