% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapPlotWindow.R
\name{overlapPlotWindow}
\alias{overlapPlotWindow}
\title{Plot overlap or excess within specified window.}
\usage{
overlapPlotWindow(A, B, t0 = 0, t1 = 24, adjust = 0.8, kmax = 3,
  type = c("overlap", "excess"), mark.window = T, n.grid = 1000, ...)
}
\arguments{
\item{A}{Numeric vectors of sighting times in [0,2pi]}

\item{B}{Numeric vectors of sighting times in [0,2pi]}

\item{t0}{The window over which to estimate the overlap and excess stats}

\item{t1}{The window over which to estimate the overlap and excess stats}

\item{adjust}{See \code{\link{overlapEst}}.}

\item{kmax}{See \code{\link{overlapEst}}.}

\item{type}{Character string with \code{overlap} to plot the overlap, 
\code{excess} to plot the area between the densities. Partial matching okay.}

\item{mark.window}{Logical indicating whether to plot vertical lines at the
window boundaries.}
}
\value{
None. Plots to the current device.
}
\description{
Densities are plotted for the entire day, but the overlap or excess are only
shown within the requested window.
}
