% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.overlapPermObj.R
\name{print.overlapPermObj}
\alias{print.overlapPermObj}
\title{S3 method print method for overlapPermObj}
\usage{
\method{print}{overlapPermObj}(obj, digits = 3)
}
\arguments{
\item{obj}{An object output by \code{\link{overlapPerm}}.}

\item{obj}{The number of digits to display in the printed table.}
}
\description{
Tidily print the list created by \code{\link{overlapPerm}}.
}
