% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapPerm.R
\name{overlapPerm}
\alias{overlapPerm}
\title{Overlap window permutation tests}
\usage{
overlapPerm(A, B, t0 = 0, t1 = 24, adjust = 0.8, nperm = 1000,
  kmax = 3, parametric = T, two.sided = T)
}
\description{
Overlap window permutation tests
}
